Added new methods involving paths, fixed issues with findFSO
Added method overloading to the methods for creating files
and directories. Now, files and directories can be created by path.
This should be easier and more useful than having to find a reference
first. Also added a new method for changing the current directory,
eventually this should be moved to be a shell member, but for now
this works.