Add new methods for creating directories and their unit tests
Added a new method override for createDirectory that accepts only
a path. This is more in line with the actual usage of 'mkdir' which
takes either just a directory name or a path with the directory name
at the end. Also added some minor methods like a getter for a
directory's children.

There are currently several unit tests for the new createDirectory
method, however, it most likely needs more unit tests as it is
possibly saucy.