FSAPI now a singleton, add some bash commands, FSAPI now in main
FSAPI's design pattern has been changed to be a Singleton, this was
actually done in the last commit, however I forgor to mention it > // <.
Now a reference to FSAPI can be obtained by calling the static method
`FileSystemAPI.getInstance()`.

FSAPI has been implemented into main. A few commands were also
implemented in order to test out the functionality of FSAPI. Namely
these commands are `cd`, `pwd`, and `mkdir`. Going forward, I will be
implementing bash commands in a different branch, as this one is only
concerned with the file system.