Add new helpers, refactor moveCurrentDirectory,
When using getFileSystemObject, I found that having to
use ``instanceof`` every time to assert that it was either a directory
or file was a little clunky. I added getFileSystemDirectory and
getFileSystemFile, that return an Optional of their respective
types -- this Optional will be empty when either the object doesn't
exist, or it is of the incorrect type. This addition to the API helped
simplify the logic for moveCurrentDirectory quite a bit.

I also added ENV sets so other packages know what PWD is.