Overhaul of IO system.
From main, you now run commands by creating a new process and
executing a command. This mirrors how bash would do it.

There are some nuances to this, but programming new commands
mostly stays the same, with one exception -- I/O is now handled
through file descriptors. The long and short of this is to
just use the new `cmdPrintln` and `cmdErrln` functions for
printing diagnostics and errors.

A README.md with a programmer's guide has been added to the
`commands/` directory -- go ahead and check that out.