Add Kernel functionality and a boatload of other changes.
I apologize in advance for the sheer size of the diff that
this commit is about to give.

The kernel now exists, and is responsable for dealing with
certain syscalls and spawning processes.

Processes exist, and they run commands. They have their own
environment, and _should_ get their own file descriptor set
down the line -- that has yet to be implemented.

JBash is now just a lowly command (but all its functionality
should still work!)

I'm sure there are other changes in this monster diff that I
forgot to address. I will probably start writing a wiki for
a lot of this stuff so we can refer back to it.